requirements.txt
laspy>=2.0.0
numpy==1.26.4
open3d>=0.15.0
Pillow>=10.0.0